refactor: create shared/db module
