+++ apps/user/src/lib/server/auth.ts
@@ -1,9 +1,9 @@
 import type { RequestEvent } from '@sveltejs/kit';
 import { deleteCookie, setCookie } from '$lib/server/cookie';
-import { db } from '$lib/server/db';
-import * as table from '$lib/server/db/schema';
 import { sha256 } from '@oslojs/crypto/sha2';
 import { encodeBase64url, encodeHexLowerCase } from '@oslojs/encoding';
+import { db } from '@shared/db';
+import * as table from '@shared/db/schema';
 import { eq } from 'drizzle-orm';
 
 const DAY_IN_MS = 1000 * 60 * 60 * 24;
