+++ shared/db/.env.example
@@ -1,0 +1,2 @@
+DATABASE_URL="./data"
+DRIVER="pglite"
