+++ shared/db/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@shared/db",
+  "type": "module",
+  "version": "0.0.1",
+  "main": "index.js",
+  "scripts": {
+    "format": "eslint --cache . --fix",
+    "lint": "eslint --cache .",
+    "push": "drizzle-kit push",
+    "migrate": "drizzle-kit migrate",
+    "studio": "drizzle-kit studio"
+  },
+  "dependencies": {
+    "@paralleldrive/cuid2": "^2.2.2",
+    "drizzle-orm": "^0.37.0"
+  },
+  "devDependencies": {
+    "@electric-sql/pglite": "^0.2.14",
+    "drizzle-kit": "^0.29.1"
+  }
+}
